Update engagement_and_scoping to fit structure
Commissioner responsibility changed to Approver  "The Approver should
ensure that there is sufficient governance in place to support the
analysis and their role in the wider project or programme. This is
particularly important if the analysis supports business critical
decisions. This  may need to be revisited at the design stage if a novel
or riskier approach is required (for example if AI models are used)."

Responsilbilities changed to use imperative keywords